minor makefile changes
- linking every tool completely statically, objview and wed are the
  only tools which need dynamic libc, libpthread, libdl and libGLU
- another small intptr_t fix, WED_VertexTool caused a crash because
  of a cast in ControlHandlyBy() (inherited from WED_HandleToolBase)

+++ src/WEDMap/WED_HandleToolBase.h
@@ -111,7 +111,7 @@ private:
 		int						mSelX;
 		int						mSelY;
 
-		int						mHandleEntity;		// Which entity do we drag
+		intptr_t				mHandleEntity;		// Which entity do we drag
 		int						mHandleIndex;
 		Point2					mTrackPoint;
